src/data_acquisition/social_media/requirements.txt
twikit==2.2.1
iso639-lang==2.5.1